source/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain-core>=0.3.10
langchain-community>=0.3.10
langchain-chroma>=0.1.2
chromadb>=0.5.4
sentence-transformers>=3.1.0
torch>=2.3.0
ibm-cloud-sdk-core>=3.20.0
requests>=2.32.3
python-dotenv>=1.0.1
pydantic>=2.9.0
